+++ sdk/storage/storage-file-share/src/Clients.ts
@@ -97,7 +97,7 @@ import {
   ListHandlesResponse as GeneratedListHandlesResponse,
 } from "./generated/src/models";
 import { Share, Directory, File } from "./generated/src/operationsInterfaces";
-import { newPipeline, Pipeline, PipelineLike } from "../../storage-blob/src/Pipeline";
+import { newPipeline, Pipeline, PipelineLike } from "./Pipeline";
 import {
   DEFAULT_MAX_DOWNLOAD_RETRY_REQUESTS,
   DEFAULT_HIGH_LEVEL_CONCURRENCY,
@@ -601,14 +601,7 @@ export class ShareClient extends StorageClient {
    */
   constructor(
     url: string,
-    credential?: StorageSharedKeyCredential | AnonymousCredential | TokenCredential,
-    // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
-    /* eslint-disable-next-line @azure/azure-sdk/ts-naming-options */
-    options?: ShareClientOptions,
-  );
-  constructor(
-    url: string,
-    credential?: StorageSharedKeyCredential | AnonymousCredential,
+    credential?: Credential | TokenCredential,
     // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
     /* eslint-disable-next-line @azure/azure-sdk/ts-naming-options */
     options?: ShareClientOptions,
@@ -627,8 +620,7 @@ export class ShareClient extends StorageClient {
   constructor(
     urlOrConnectionString: string,
     credentialOrPipelineOrShareName?:
-      | StorageSharedKeyCredential
-      | AnonymousCredential
+      | Credential
       | TokenCredential
       | PipelineLike
       | string,
@@ -1629,7 +1621,7 @@ export class ShareDirectoryClient extends StorageClient {
    */
   constructor(
     url: string,
-    credential?: AnonymousCredential | StorageSharedKeyCredential | TokenCredential,
+    credential?: Credential | TokenCredential,
     // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
     /* eslint-disable-next-line @azure/azure-sdk/ts-naming-options */
     options?: ShareClientOptions,
@@ -1652,8 +1644,7 @@ export class ShareDirectoryClient extends StorageClient {
   constructor(
     url: string,
     credentialOrPipeline?:
-      | AnonymousCredential
-      | StorageSharedKeyCredential
+      | Credential
       | TokenCredential
       | Pipeline,
     // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
@@ -2937,6 +2928,13 @@ export interface FileGetRangeListOptions extends CommonOptions {
    * Lease access conditions.
    */
   leaseAccessConditions?: LeaseAccessConditions;
+  /**
+   * This header is allowed only when prevShareSnapshot parameter is set.
+   * Determines whether the changed ranges for a file that has been renamed or moved between the target snapshot (or the live file) and the previous snapshot should be listed.
+   * If the value is true, the valid changed ranges for the file will be returned. If the value is false, the operation will result in a failure with 409 (Conflict) response.
+   * The default value is false.
+   */
+  includeRenames?: boolean;
 }
 
 /**
@@ -3504,7 +3502,7 @@ export class ShareFileClient extends StorageClient {
    */
   constructor(
     url: string,
-    credential?: AnonymousCredential | StorageSharedKeyCredential | TokenCredential,
+    credential?: Credential | TokenCredential,
     // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
     /* eslint-disable-next-line @azure/azure-sdk/ts-naming-options */
     options?: ShareClientOptions,
@@ -3527,8 +3525,7 @@ export class ShareFileClient extends StorageClient {
   constructor(
     url: string,
     credentialOrPipeline?:
-      | AnonymousCredential
-      | StorageSharedKeyCredential
+      | Credential
       | TokenCredential
       | Pipeline,
     // Legacy, no way to fix the eslint error without breaking. Disable the rule for this line.
@@ -4225,6 +4222,7 @@ export class ShareFileClient extends StorageClient {
         >(
           await this.context.getRangeList({
             ...updatedOptions,
+            supportRename: updatedOptions.includeRenames,
             prevsharesnapshot: prevShareSnapshot,
             range: updatedOptions.range ? rangeToString(updatedOptions.range) : undefined,
             ...this.shareClientConfig,