+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -4,11 +4,10 @@
 
 ### Features Added
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
+- Added support for service version 2024-05-04.
+- Added ability to call ShareFileClient.getRangeListDiff() on a file that has been renamed.
+- Added support OAuth for interface ShareFileClient.uploadRangeFromURL().
+- Added support to return clientName property in file/directory handle properties in ShareFileClient.listHandles() and ShareDirectoryClient.listHandles().
 
 - Allow HTTP connections
 
@@ -17,6 +16,28 @@
 ### Other Changes
 
 - Migrated dependency on `@azure/core-http` to `@azure/core-rest-pipeline`.
+
+## 12.17.0 (2023-11-09)
+
+### Features Added
+
+- Includes all features released in 12.17.0-beta.1.
+
+## 12.17.0-beta.1 (2023-10-18)
+
+### Features Added
+
+- Added support for service version 2023-11-03.
+- Added support for indicating audience in StoragePipelineOptions.
+
+## 12.16.1 (Unreleased)
+
+### Features Added
+
+- Added support for service version 2023-11-03.
+- Added support for indicating audience in StoragePipelineOptions.
+
+## 12.16.0 (2023-09-14)
 
 ## 12.15.0 (2023-07-12)